requirements.txt
# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning and clustering
scikit-learn>=1.0.0

# Deep learning
torch>=1.9.0
torchvision>=0.10.0

# Image processing
Pillow>=8.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# COCO dataset tools
pycocotools>=2.0.4

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.62.0
requests>=2.25.0

# Text processing (if needed for advanced text features)
nltk>=3.6.0

# Additional utilities for Python < 3.8 compatibility
# Note: pathlib2 is not needed for Python 3.8+, using built-in pathlib 